requirements.txt
pyserial
pandas
dash
plotly
matplotlib
pyqtgraph
pyqt5
pyyaml
numpy
